+++ src/apigateway/error/index.ts
@@ -4,3 +4,4 @@
 
 export * from './api-error';
 export * from './api-timeout';
+export * from './build-path-not-found';