+++ src/index.ts
@@ -15,9 +15,7 @@ export {
     RouteResolver,
     ResolverCache,
     ImportManager,
-    PatternResolver,
-    DirectoryResolver,
-    ListResolver
+    PatternResolver
 } from './apigateway/resolver';
 
 // Common utilities exports